+++ src/dataset/scrobble.js
@@ -1,0 +1,25 @@
+export function uncompress([
+  date,
+  trackName,
+  trackPlaycount,
+  albumName,
+  albumPlaycount,
+  artistName,
+  artistPlaycount,
+]) {
+  return {
+    date,
+    track: {
+      name: trackName,
+      playcount: trackPlaycount,
+    },
+    album: {
+      name: albumName,
+      playcount: albumPlaycount,
+    },
+    artist: {
+      name: artistName,
+      playcount: artistPlaycount,
+    },
+  };
+}